Added support for lightbox on wordpress galleries *experimental*
Replaces [gallery] shortcode with an almost identical copy from the
WP trunk code (after 3.2.1).

Adjustments find and preg_replace the rel=lightbox and title
attributes into the link. I don't know of doing this without
replacing the [gallery] shortcode. But it does mean that this has
to be kept up to date with any changes to WP core.